tests/wpt/tests/css/css-scroll-snap/parsing/scroll-snap-align-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-snap-align with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-snap-align">
<meta name="assert" content="scroll-snap-align supports the full grammar '[ none | start | end | center ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-snap-align", "none");
test_valid_value("scroll-snap-align", "start");
test_valid_value("scroll-snap-align", "end");
test_valid_value("scroll-snap-align", "center");

test_valid_value("scroll-snap-align", "start none");
test_valid_value("scroll-snap-align", "center end");
test_valid_value("scroll-snap-align", "start start", "start");

</script>
</body>
</html>
